requirements.txt
asgiref==3.5.2
cycler==0.11.0
Django==4.1
fonttools==4.34.4
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.2
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
